% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Calculates the mean of a sparse numeric vector}
\usage{
\S4method{mean}{sparse_numeric}(x)
}
\arguments{
\item{x}{Sparse numeric vector}
}
\value{
Resulting mean of the sparse numeric vector
}
\description{
Method that computes the mean of a sparse numeric vector, by summing up the non-zero elements and dividing by the length of the vector, returning a number that is the mean
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 9L), length = 10L)
mean(x)
}
