% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{norm,sparse_numeric-method}
\alias{norm,sparse_numeric-method}
\title{Norm of a sparse numeric vector}
\usage{
\S4method{norm}{sparse_numeric}(x)
}
\arguments{
\item{x}{Sparse Numeric Vector}
}
\value{
Resulting norm
}
\description{
Method that computes the norm of the entire vector, by squaring each non-zero x value and then summing them together, and afterwards performing the squareroot, which results in the norm of the entire vector
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 9L), length = 10L)
norm(x)
}
