% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{norm}
\alias{norm}
\title{Norm of a sparse numeric vector}
\usage{
norm(x)
}
\arguments{
\item{x}{Sparse Numeric Vector}
}
\value{
Resulting norm
}
\description{
Generic function that returns the norm of the entire vector
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 9L), length = 10L)
norm(x)
}
