% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot sparse numeric vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{Sparse numeric vector}

\item{y}{Sparse numeric vector}
}
\value{
Plotted sparse numeric vectors
}
\description{
Method that plots the information of two sparse numeric vectors, including the non-zeros values, their positions, and the length of the vector with the positions on the x-axis, and non-zeros values on the y-axis, each vector represented by a different color
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 7L), length = 10L)
y <- new("sparse_numeric", value = c(3,8), pos=c(5L, 7L), length = 10L)
plot(x,y)
}
