% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{show,sparse_numeric-method}
\alias{show,sparse_numeric-method}
\title{Show a sparse numeric vector}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{Sparse numeric vector}
}
\value{
Printed sparse numeric vector parameters
}
\description{
Method that prints the information of a sparse numeric vector, including the non-zeros values, their positions, and the length of the vector
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 9L), length = 10L)
x
}
