% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_add}
\alias{sparse_add}
\title{Add two sparse vectors together}
\usage{
sparse_add(x, y)
}
\arguments{
\item{x}{Sparse Numeric Vector}

\item{y}{Sparse Numeric Vector}
}
\value{
Sparse numeric vector as a sum of the two parameters
}
\description{
Generic function that adds two sparse numeric vectors, into one resulting vector
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 9L), length = 10L)
y <- new("sparse_numeric", value = c(3,8), pos=c(1L, 7L), length = 10L)
sparse_add(x,y)
}
