% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Cross product of two sparse vectors}
\usage{
\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{Sparse Numeric Vector}

\item{y}{Sparse Numeric Vector}
}
\value{
Resulting cross product
}
\description{
Method that computes the cross product of two sparse numeric vectors, by merging them and performing multiplication on non-zeroes and then summing them all up, returning one number that is the cross product
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 9L), length = 10L)
y <- new("sparse_numeric", value = c(3,8), pos=c(1L, 7L), length = 10L)
sparse_crossprod(x,y)
}
