% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\title{Cross product of two sparse vectors}
\usage{
sparse_crossprod(x, y)
}
\arguments{
\item{x}{Sparse Numeric Vector}

\item{y}{Sparse Numeric Vector}
}
\value{
Resulting cross product
}
\description{
Generic function that returns the cross product of two sparse numeric vectors, into one resulting number
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 9L), length = 10L)
y <- new("sparse_numeric", value = c(3,8), pos=c(1L, 7L), length = 10L)
sparse_crossprod(x,y)
}
