% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_div,sparse_numeric,sparse_numeric-method}
\alias{sparse_div,sparse_numeric,sparse_numeric-method}
\title{Divides one sparse vector by another}
\usage{
\S4method{sparse_div}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{Sparse numeric vector}

\item{y}{Sparse numeric vector}
}
\value{
Sparse numeric vector as a quotient of the two parameters
}
\description{
Method that divides one sparse numeric vector by another, by merging them and performing division on non-zeroes, returning a vector that is the quotient
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 7L), length = 10L)
y <- new("sparse_numeric", value = c(3,8), pos=c(5L, 7L), length = 10L)
sparse_div(x,y)
}
