% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_div}
\alias{sparse_div}
\title{Division of two sparse vectors}
\usage{
sparse_div(x, y)
}
\arguments{
\item{x}{Sparse Numeric Vector}

\item{y}{Sparse Numeric Vector}
}
\value{
Sparse numeric vector as a quotient of the two parameters
}
\description{
Generic function that returns the quotient of two sparse numeric vectors, into one resulting vector
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 7L), length = 10L)
y <- new("sparse_numeric", value = c(3,8), pos=c(5L, 7L), length = 10L)
sparse_div(x,y)
}
