% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Multiplies two sparse vectors together}
\usage{
\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{Sparse numeric vector}

\item{y}{Sparse numeric vector}
}
\value{
Sparse numeric vector as a product of the two parameters
}
\description{
Method that multiplies two sparse numeric vectors, by merging them and performing multiplication on non-zeroes, returning a vector that is the product
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 7L), length = 10L)
y <- new("sparse_numeric", value = c(3,8), pos=c(5L, 7L), length = 10L)
sparse_mult(x,y)
}
