% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Define a sparse numeric vector class}
\value{
Sparse numeric vector
}
\description{
Define a sparse numeric vector class, meaning that it stores non-zero numeric values in a vector, their given positions, and the length of the vector.
}
\section{Slots}{

\describe{
\item{\code{value}}{The non zero numbers in the vector}

\item{\code{pos}}{The position/index of these non zeroes numbers in the vector}

\item{\code{length}}{The total length of the vector}
}}

\examples{
x <- new("sparse_numeric", value = c(4,3.2, 6.1), pos=c(2L, 3L, 5L), length = 50L)
}
