% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{-,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\title{Subtract two sparse vectors together using -}
\usage{
\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{Sparse numeric vector}

\item{e2}{Sparse numeric vector}
}
\value{
Sparse numeric vector as a difference of the two parameters
}
\description{
Method that performs subtraction with - operator by calling sparse_sub
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 9L), length = 10L)
y <- new("sparse_numeric", value = c(3,8), pos=c(1L, 7L), length = 10L)
z <- x - y
z
}
