% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_numeric_setas_n}
\alias{sparse_numeric_setas_n}
\title{Convert sparse to numeric}
\arguments{
\item{from}{Sparse Numeric vector}
}
\value{
Numeric vector
}
\description{
Convert sparse numeric vector format into numeric vector
}
\examples{
x <- new("sparse_numeric", value = c(4,3.2, 6.1), pos=c(2L, 3L, 5L), length = 50L)
as(x, "numeric")

}
