% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_numeric_setas_s}
\alias{sparse_numeric_setas_s}
\title{Convert numeric to sparse}
\arguments{
\item{from}{Numeric vector}
}
\value{
Sparse numeric vector
}
\description{
Convert numeric vector into sparse numeric vector format
}
\examples{
x <- c(0,0,0,0,3,0,0)
as(x, "sparse_numeric")

}
