% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_numeric_validity}
\alias{sparse_numeric_validity}
\title{Check the validity of the sparse numeric vector}
\description{
Ensure that the sparse numeric vector, only contains numbers, does not only contain zeroes, and has an appropriate set length
}
