% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtracts two sparse vectors together}
\usage{
\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{Sparse numeric vector}

\item{y}{Sparse numeric vector}
}
\value{
Sparse numeric vector as a difference of the two parameters
}
\description{
Method that subtracts two sparse numeric vectors, by merging them and performing subtraction on non-zeroes, returning a vector that is the difference
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 9L), length = 10L)
y <- new("sparse_numeric", value = c(3,8), pos=c(1L, 7L), length = 10L)
sparse_sub(x,y)
}
