% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{standardize,sparse_numeric-method}
\alias{standardize,sparse_numeric-method}
\title{Standardizes a sparse numeric vector}
\usage{
\S4method{standardize}{sparse_numeric}(x)
}
\arguments{
\item{x}{Sparse Numeric Vector}
}
\value{
Standardized Sparse Numeric Vector
}
\description{
Method that standardizes the entire vector, by taking each element of the vector and subtracting off the vector mean and dividing by the vector's standard deviation
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 9L), length = 10L)
standardize(x)
}
