% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{standardize}
\alias{standardize}
\title{Standardizes a sparse numeric vector}
\usage{
standardize(x)
}
\arguments{
\item{x}{Sparse Numeric Vector}
}
\value{
Standardized Sparse Numeric Vector
}
\description{
Generic function that standardizes the entire vector
}
\examples{
x <- new("sparse_numeric", value = c(5,7), pos=c(5L, 9L), length = 10L)
standardize(x)
}
